refactor: extract SrLevelsCard component and add coverage
